requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
APScheduler
requests
beautifulsoup4
python-dotenv